+++ packages/eyes-sdk-core/test/unit/PreserveTargetFrameAfterCheck.spec.js
@@ -1,7 +1,9 @@
 const assert = require('assert')
 const {startFakeEyesServer} = require('@applitools/sdk-fake-eyes-server')
-const MockDriver = require('../utils/MockDriver')
+const snippets = require('@applitools/snippets')
+const {MockDriver} = require('@applitools/driver')
 const {EyesClassic} = require('../utils/FakeSDK')
+const {generateScreenshot} = require('../utils/FakeScreenshot')
 
 describe('PreserveCheckSettingsFrameAfterCheck', () => {
   let server, serverUrl, driver, eyes
@@ -12,6 +14,10 @@ describe('PreserveCheckSettingsFrameAfterCheck', () => {
 
   before(async () => {
     driver = new MockDriver()
+    driver.takeScreenshot = generateScreenshot
+    driver.mockScript(snippets.getElementContentSize, ([element]) => {
+      return element.rect || {x: 0, y: 0, width: 100, height: 100}
+    })
     driver.mockElements([
       {
         selector: 'frame1',