+++ packages/eyes-sdk-core/test/utils/FakeSDK.js
@@ -1,10 +1,10 @@
-const spec = require('./FakeSpecDriver')
-const {EyesSDK} = require('../../index')
+const {fake} = require('@applitools/driver')
 const VisualGridClient = require('@applitools/visual-grid-client')
+const {EyesSDK} = require('../../index')
 
 module.exports = EyesSDK({
   name: 'eyes.fake',
   version: '0.2.0',
-  spec,
+  spec: fake.spec,
   VisualGridClient,
 })